Migrate PCI Root Bridge HOB to Universal Payload specification
This commit migrates the PCI Root Bridge information from the legacy
gLoaderPciRootBridgeInfoGuid to the Universal Payload specification
using gUniversalPayloadPciRootBridgeInfoGuid.

Key changes:
- Remove legacy PciRootBridgeInfoGuid.h and update BootloaderCommonPkg.dec
- Update PciEnumerationLib to use Universal Payload PCI root bridge structures
- Add BuildUniversalPayloadPciRootBridgeHob() function implementation
- Update all platform Stage2BoardInitLib headers to use new Universal Payload includes
- Fix type casting issues in Idaville platform for ACPI table compatibility
- Update PayloadEntryLib to consume Universal Payload PCI root bridge HOBs
- Update VTdLib to use new Universal Payload GUID

Platforms verified:
- QEMU: Build successful
- Idaville: Build successful with type casting fixes
- ArrowLake: Build successful
- All other platforms: Headers updated for compatibility

This migration maintains backward compatibility while adopting the
Universal Payload standard for better interoperability.

+++ Platform/IdavilleBoardPkg/Library/Stage2BoardInitLib/Stage2BoardInitLib.c
@@ -1209,8 +1209,8 @@ PlatformUpdateAcpiGnvs (
           AcpiParameter->Io16Length[Index]    = (UINT16) (ResRange->IoLimit - ResRange->IoBase + 1);
         }
         if (ResRange->Mmio32Limit != 0) {
-          AcpiParameter->Mmio32Base[Index]    = ResRange->Mmio32Base;
-          AcpiParameter->Mmio32Length[Index]  = (ResRange->Mmio32Limit - ResRange->Mmio32Base + 1);
+          AcpiParameter->Mmio32Base[Index]    = (UINT32)ResRange->Mmio32Base;
+          AcpiParameter->Mmio32Length[Index]  = (UINT32)(ResRange->Mmio32Limit - ResRange->Mmio32Base + 1);
         }
         if (ResRange->Mmio64Limit != 0) {
           AcpiParameter->Mmio64Base[Index]    = ResRange->Mmio64Base;
